+++ src/mainworker.h
@@ -12,17 +12,10 @@
 extern "C" {
 #endif
 
-enum wsim_end_mode_t {
-  WSIM_END_NORMAL,
-  WSIM_END_SIGNAL,
-  WSIM_END_ERROR
-};
-
 /*
  * Start simulation main loop.
  */
-int startWorker(int argc, char* argv[]);
-void main_end(enum wsim_end_mode_t mode);
+int simulationMain(int argc, char* argv[]);
 
 #ifdef __cplusplus
 }
